CLI-023, CLI-024: tail and management subtree explicit dispatch
Add explicit Command::Tail and Command::Management dispatch arms
in execute_command() with stub_not_implemented outputs.

Integration tests: tail_bare, tail_token, management_bare,
management_token — all verify parse-dispatch round-trip.

Extract inline test module from main.rs to main_tests.rs to
reduce debtmap GodObject score (102 methods → 24 production).

Ledger: CLI-023 and CLI-024 rows updated to baseline-backed
tests; remaining work is behavioral (CDC-026 WebSocket
transport, CDC-038 management token API).
