+++ docs/parity/cli/implementation-checklist.md
@@ -111,8 +111,8 @@ hand-written parser (no clap or structopt).
 | ID | Feature group | Baseline source | Baseline behavior or contract | Rust owner now | Rust status now | Parity evidence status | Divergence status | Required tests | Priority | Notes |
 | --- | --- | --- | --- | --- | --- | --- | --- | --- | --- | --- |
 | CLI-022 | access subtree | `cmd/cloudflared/access/cmd.go` | `access` command (alias `forward`) with subcommands: `login` (browser auth), `curl` (JWT injection), `token` (JWT production), `tcp` (aliases `rdp`, `ssh`, `smb` for TCP/RDP/SSH/SMB proxy), `ssh-config` (print SSH config), `ssh-gen` (generate short-lived cert); `--fedramp` flag | current CLI surface | audited, partial | baseline-backed tests | open gap | per-subcommand behavior tests | high | CLI surface parity-backed: bare `access` and `forward` alias both show access help (exit 0), all 6 subcommands + 3 tcp aliases parsed and dispatched explicitly; 6 parse-dispatch tests in `cfdrs-cli`, 12 integration tests in `cfdrs-bin` covering help, alias equivalence, and per-subcommand dispatch; behavioral implementations are stubs pending CDC/HIS runtime infrastructure (browser auth, TCP proxy, JWT, SSH certs). Go baseline has no `NArg` constraints on access subcommands. See [docs/parity/cli/access-subtree.md](access-subtree.md) |
-| CLI-023 | tail subtree | `cmd/cloudflared/tail/cmd.go` | `tail [TUNNEL-ID]` streams remote logs; flags: `--connector-id`, `--event` (filter: cloudflared/http/tcp/udp), `--level` (default `debug`), `--sample` (default 1.0), `--token` (env `TUNNEL_MANAGEMENT_TOKEN`), `--management-hostname` (hidden, default `management.argotunnel.com`), `--trace` (hidden), `--loglevel` (default `info`, env `TUNNEL_LOGLEVEL`), `--origincert` (env `TUNNEL_ORIGIN_CERT`), `--output` (json/default); hidden subcommand `token` gets management JWT | current CLI surface | audited, partial | local tests | open gap | help crawl, filter tests, hidden `token` subcommand tests, output-format tests | high | Rust parses tail and hidden `token` subcommand; dispatches to stubs; 2 parse-dispatch tests (`tail_bare`, `tail_token`). CDC owns log-streaming contract. See [docs/parity/cli/tail-and-management.md](tail-and-management.md) |
-| CLI-024 | management subtree | `cmd/cloudflared/management/cmd.go` | `management` (hidden, category `Management`) with hidden subcommand `token`; token subcommand requires `--resource` (values: `logs`, `admin`, `host_details`), `--origincert`, `--loglevel`, `--output` (json/default) | current CLI surface | audited, partial | local tests | open gap | hidden-command help capture, token invocation tests, resource-flag tests | medium | Rust parses management and hidden `token` subcommand; dispatches to stubs; 2 parse-dispatch tests (`management_bare`, `management_token`) |
+| CLI-023 | tail subtree | `cmd/cloudflared/tail/cmd.go` | `tail [TUNNEL-ID]` streams remote logs; flags: `--connector-id`, `--event` (filter: cloudflared/http/tcp/udp), `--level` (default `debug`), `--sample` (default 1.0), `--token` (env `TUNNEL_MANAGEMENT_TOKEN`), `--management-hostname` (hidden, default `management.argotunnel.com`), `--trace` (hidden), `--loglevel` (default `info`, env `TUNNEL_LOGLEVEL`), `--origincert` (env `TUNNEL_ORIGIN_CERT`), `--output` (json/default); hidden subcommand `token` gets management JWT | current CLI surface | audited, partial | baseline-backed tests | open gap | WebSocket transport tests, management token integration tests, output-format tests | high | CLI surface parity-backed: bare `tail` and hidden `token` subcommand parsed and dispatched explicitly; 2 parse-dispatch tests in `cfdrs-cli`, 4 integration tests in `cfdrs-bin` (`tail_bare_dispatches_to_stub`, `tail_token_dispatches_to_stub`, `management_bare_dispatches_to_stub`, `management_token_dispatches_to_stub`); behavioral implementations are stubs pending CDC-026 WebSocket transport and CDC-038 management token API. CDC owns log-streaming contract. See [docs/parity/cli/tail-and-management.md](tail-and-management.md) |
+| CLI-024 | management subtree | `cmd/cloudflared/management/cmd.go` | `management` (hidden, category `Management`) with hidden subcommand `token`; token subcommand requires `--resource` (values: `logs`, `admin`, `host_details`), `--origincert`, `--loglevel`, `--output` (json/default) | current CLI surface | audited, partial | baseline-backed tests | open gap | management token invocation tests, resource-flag tests | medium | CLI surface parity-backed: hidden `management` command and hidden `token` subcommand parsed and dispatched explicitly; 2 parse-dispatch tests in `cfdrs-cli`, 4 integration tests shared with CLI-023 in `cfdrs-bin`; behavioral implementations are stubs pending CDC-038 management token API. See [docs/parity/cli/tail-and-management.md](tail-and-management.md) |
 
 ### Compatibility, Formatting, And Error Behavior
 
@@ -215,14 +215,14 @@ High gaps (behavioral implementation behind stubs):
 
 - CLI-006: update command (parsed — needs HIS updater)
 - CLI-022: access subtree (CLI surface parity-backed — behavioral stubs pending CDC/HIS runtime)
-- CLI-023: tail subtree (hidden token subcommand parsed, dispatched to stub)
+- CLI-023: tail subtree (CLI surface parity-backed — behavioral stubs pending CDC-026 WebSocket transport and CDC-038 management token API)
 
 Medium gaps (all now parsed and dispatched to stubs):
 
 - CLI-017: tunnel ready (depends on HIS-024/025)
 - CLI-018: tunnel diag (depends on HIS diagnostics)
 - CLI-021: tunnel ingress (hidden, parsed with validate/rule nesting)
-- CLI-024: management subtree (hidden, token subcommand parsed)
+- CLI-024: management subtree (CLI surface parity-backed — behavioral stubs pending CDC-038 management token API)
 
 ## Scope Classification
